update minimum node version to 10.15update dependencies, minor code changes
