+++ gulp/tasks/01-custom-html-templates.js
@@ -36,6 +36,6 @@ function prepareTemplates() {
     }
 }
 
-gulp.task('custom-html-templates', ['select-view'], () => {
+gulp.task('custom-html-templates', gulp.series('select-view', () => {
     prepareTemplates();
-})
+}))
