+++ gulp/tasks/04-custom-css.js
@@ -11,19 +11,19 @@ var glob = require('glob');
 
 let buildParams = config.buildParams;
 
-gulp.task('watch-css', ['select-view'], () => {
+gulp.task('watch-css', gulp.series('select-view', () => {
 
     gulp.watch([buildParams.customCssMainPath(),buildParams.customNpmCssPath(),'!'+buildParams.customCssPath()],['custom-css']);
-});
+}));
 
 
 
 
-gulp.task('custom-css', ['select-view'], () => {
+gulp.task('custom-css', gulp.series('select-view', () => {
 
     return gulp.src([buildParams.customCssMainPath(),buildParams.customNpmCssPath(),'!'+buildParams.customCssPath()])
         .pipe(concat(buildParams.customCssFile))
         .pipe(gulp.dest(buildParams.viewCssDir()));
 
 
-});
+}));
