+++ gulp/tasks/09-images.js
