+++ package.json
@@ -4,59 +4,64 @@
   "description": "The Exlibris Primo Open Discovery Framework",
   "author": "noamamit92",
   "devDependencies": {
+    "babel-core": "6.26.3",
     "babel-plugin-transform-html-import-to-string": "0.0.1",
-    "babel-preset-es2015": "6.6.0",
-    "babelify": "7.3.0",
-    "bluebird": "3.3.1",
-    "browser-sync": "2.9.9",
-    "browserify": "13.1.1",
+    "babel-preset-env": "1.7.0",
+    "babelify": "8.0.0",
+    "bluebird": "3.5.4",
+    "browser-sync": "^2.26.5",
+    "browserify": "16.2.3",
     "camel-case": "^3.0.0",
-    "colors": "^1.3.1",
-    "concat-stream": "1.5.1",
-    "connect-modrewrite": "0.8.2",
-    "css": "2.2.1",
-    "css-color-extractor": "0.0.4",
-    "css-color-extractor-cli": "0.0.1",
-    "del": "2.2.1",
-    "fs-extra": "2.1.2",
-    "glob": "6.0.4",
-    "gulp": "3.9.1",
-    "gulp-angular-templatecache": "^2.0.0",
-    "gulp-autoprefixer": "3.1.0",
-    "gulp-babel": "6.1.2",
-    "gulp-clone": "1.0.0",
-    "gulp-concat": "2.6.0",
-    "gulp-cssnano": "2.1.2",
-    "gulp-debug": "2.1.2",
-    "gulp-flatten": "^0.3.1",
-    "gulp-plumber": "1.1.0",
-    "gulp-rename": "1.2.2",
-    "gulp-sass": "2.3.2",
-    "gulp-sourcemaps": "1.6.0",
+    "colors": "^1.3.3",
+    "concat-stream": "^2.0.0",
+    "connect-modrewrite": "0.10.2",
+    "css": "^2.2.4",
+    "css-color-extractor": "0.0.5",
+    "css-color-extractor-cli": "0.0.2",
+    "del": "4.1.1",
+    "fs-extra": "7.0.1",
+    "glob": "7.1.3",
+    "graceful-fs": "4.1.15",
+    "gulp": "4.0.1",
+    "gulp-angular-templatecache": "^2.2.7",
+    "gulp-autoprefixer": "6.1.0",
+    "gulp-babel": "7.0.1",
+    "gulp-clone": "2.0.1",
+    "gulp-concat": "2.6.1",
+    "gulp-cssnano": "^2.1.3",
+    "gulp-debug": "4.0.0",
+    "gulp-flatten": "^0.4.0",
+    "gulp-plumber": "1.2.0",
+    "gulp-rename": "1.4.0",
+    "gulp-sass": "4.0.2",
+    "gulp-sourcemaps": "2.6.5",
     "gulp-streamify": "1.0.2",
-    "gulp-template": "4.0.0",
-    "gulp-uglify": "3.0.0",
-    "gulp-wrap": "0.11.0",
-    "gulp-zip": "3.1.0",
-    "http-response-object": "1.1.0",
-    "listdirs": "3.0.0",
+    "gulp-template": "5.0.0",
+    "gulp-uglify": "3.0.2",
+    "gulp-util": "3.0.8",
+    "gulp-wrap": "^0.15.0",
+    "gulp-zip": "4.2.0",
+    "http-response-object": "3.0.2",
+    "listdirs": "3.1.1",
     "lodash": "4.17.11",
-    "merge-stream": "1.0.0",
+    "merge-stream": "1.0.1",
+    "minimatch": "3.0.4",
     "minimist": "1.2.0",
     "prompt": "1.0.0",
-    "protractor": "5.0.0",
-    "protractor-image-comparison": "1.2.3",
-    "request": "2.72.0",
-    "require-dir": "0.3.2",
-    "run-sequence": "1.2.1",
-    "shelljs": "0.6.0",
+    "protractor": "5.4.2",
+    "protractor-image-comparison": "3.1.0",
+    "request": "^2.88.0",
+    "require-dir": "1.2.0",
+    "run-sequence": "2.2.1",
+    "shelljs": "0.8.3",
     "stream-to-promise": "^2.2.0",
-    "tar-fs": "1.13.0",
-    "vinyl-buffer": "1.0.0",
-    "vinyl-source-stream": "1.1.0"
+    "tar-fs": "2.0.0",
+    "vinyl-buffer": "1.0.1",
+    "vinyl-source-stream": "2.0.0"
   },
   "engines": {
-    "node": ">=4.2"
+    "node": ">=10.0"
   },
-  "repository": "https://github.com/ExLibrisGroup/primo-explore-devenv"
+  "repository": "https://github.com/ExLibrisGroup/primo-explore-devenv",
+  "dependencies": {}
 }
